+++ aws-android-sdk-mobile-client/src/androidTest/java/com/amazonaws/mobile/client/AWSMobileClientTest.java
@@ -276,6 +276,17 @@ public class AWSMobileClientTest extends AWSMobileClientTestBase {
         userAttributes.put("email", EMAIL);
 
         final SignUpResult signUpResult = auth.signUp(username, PASSWORD, userAttributes, null);
+        // Check for non-null user sub in the SignUpResult
+        String userSub = signUpResult.getUserSub();
+        assertNotNull(userSub);
+        // Validate that the userSub is a valid UUID
+        assertEquals(36, userSub.length());
+        assertEquals(5, userSub.split("-").length);
+        assertEquals(8,  userSub.split("-")[0].length());
+        assertEquals(4,  userSub.split("-")[1].length());
+        assertEquals(4,  userSub.split("-")[2].length());
+        assertEquals(4,  userSub.split("-")[3].length());
+        assertEquals(12,  userSub.split("-")[4].length());
         if (signUpResult.getConfirmationState()) {
             // Done
         } else {
@@ -284,6 +295,7 @@ public class AWSMobileClientTest extends AWSMobileClientTestBase {
             assertEquals("email", details.getAttributeName());
             assertEquals("EMAIL", details.getDeliveryMedium());
         }
+        assertNotNull(signUpResult.getUserSub());
         final SignInResult signInResult = auth.signIn(username, PASSWORD, null);
         assertEquals("Cannot support MFA in tests", SignInState.DONE, signInResult.getSignInState());
     }