+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Change Log - AWS SDK for Android
 
+## [Release 2.16.0](https://github.com/aws/aws-sdk-android/releases/tag/release_v2.16.0)
+
+### New Features
+
+- **AWS Mobile Client**
+  - **Breaking API Change** `SignUpResult` available in the user callback provided during sign up now contains UserSub(UID).
+
+- **Amazon CognitoIdentityProvider**
+  - **Breaking API Change** `SignUpHandler` now receives `SignUpResult` upon success instead of a ConfirmationState and `CognitoUserCodeDeliveryDetails`
+  - `signUpInBackground` and `signUp` now receive `SignUpResult` in the `onSuccess` callback
+
 ## [Release 2.15.2](https://github.com/aws/aws-sdk-android/releases/tag/release_v2.15.2)
 
 ### Bug Fixes